# Apply relevant label matchers in LabelValues before fetching extra postings