@roidelapluie 2023-04-20
Do we need a test to make sure that `__name__` applies multiples times, like in `{__name__="up",__name__="ALERTS"}`, still works? It also feels like  we can have a `break` at the end of the for