@Logiraptor 2023-04-24
@roidelapluie Sorry if I am misunderstanding - a query with multiple `__name__` equality matchers should match 0 series, correct? Since each matcher is `and`'d together, no series can have both names. In that case, the code here would first filter out everything that isn't `up`, then everything that isn't `ALERTS`, leaving an empty list of values. I believe that's correct, and adding a `break` to the for loop would mean we'd return `["up"]` instead of an empty list.

Am I missing something?